// commerce-operations-app/e2e/enums/inLineErrors.ts
export enum ErrorFields {
  SIGNIN_EMAIL = "signin-email",
  SIGNIN_PASSWORD = "signin-password",

  //SIGNUP
  INVALID_CREDENTIALS="auth-message",
  SIGNUP_FULLNAME = "full-name",
  SIGNUP_WORK_EMAIL = "work-email",
  SIGNUP_COMPANY_NAME = "company-name",
  SIGNUP_PASSWORD = "password",
  SIGNUP_CONFIRM_PASSWORD = "confirm-password",
  FORGET_PASSWORD_EMAIL = "reset-email",

  //CREATE PRODUCTS
  PRODUCT_NAME="product-name",
  PRODUCT_CODE="product-code",
  CATEGORY="category",
  PRICE="price",
  STOCK="stock",
  PRODUCT_IMAGES="product-images"
}